+++ pypgx/cli/plot_bam_copy_number.py
@@ -6,20 +6,25 @@ import fuc
 import pysam
 
 description = f"""
-Plot copy number profile with BAM data.
+Plot copy number profile from CovFrame[CopyNumber].
 """
 
 def create_parser(subparsers):
     parser = fuc.api.common._add_parser(
         subparsers,
         fuc.api.common._script_name(),
-        help='Plot copy number profile with BAM data.',
+        help='Plot copy number profile from CovFrame[CopyNumber].',
         description=description,
     )
     parser.add_argument(
         'copy_number',
         metavar='copy-number',
         help='Archive file with the semantic type CovFrame[CopyNumber].'
+    )
+    parser.add_argument(
+        '--fitted',
+        action='store_true',
+        help='Show the fitted line as well.'
     )
     parser.add_argument(
         '--path',
@@ -45,13 +50,16 @@ def create_parser(subparsers):
         help='Y-axis top.'
     )
     parser.add_argument(
-        '--fitted',
-        action='store_true',
-        help='Show the fitted line as well.'
+        '--fontsize',
+        metavar='FLOAT',
+        type=float,
+        default=25,
+        help='Text fontsize (default: 25).'
     )
 
 def main(args):
     plot.plot_bam_copy_number(
-        args.copy_number, path=args.path, samples=args.samples,
-        ymin=args.ymin, ymax=args.ymax, fitted=args.fitted
+        args.copy_number, fitted=args.fitted, path=args.path,
+        samples=args.samples, ymin=args.ymin, ymax=args.ymax,
+        fontsize=args.fontsize
     )